[유치수] [NOBTS] endTime:long to elapsed:int
git-svn-id: http://svn.bds.nhncorp.com/pe/hippo-server/trunk@915 84d0f5b1-2673-498c-a247-62c4ff18d310

+++ src/test/java/com/profiler/server/dao/hbase/TraceDaoTest.java
@@ -92,7 +92,7 @@ public class TraceDaoTest {
         UUID uuid = UUID.randomUUID();
         List<Annotation> ano = Collections.emptyList();
         long l = System.currentTimeMillis();
-        Span span = new Span("UnitTest", uuid.getMostSignificantBits(), uuid.getLeastSignificantBits(), l, l + 5, "test", "rpc", 1, ano, "protocol:ip:port", false);
+        Span span = new Span("UnitTest", uuid.getMostSignificantBits(), uuid.getLeastSignificantBits(), l, (int) l + 5, "test", "rpc", 1, ano, "protocol:ip:port", false);
         return span;
     }
 }
